refactor(trigger_webhook): refactor the code to match the coding guidlines (SCRUM-244)
